扩展rpc - 序列化器与SPI机制

+++ salmon-rpc/example-common/src/main/java/com/salmon/service/UserService.java
@@ -23,7 +23,7 @@ public interface UserService {
      *
      * @return 短整型数字
      */
-    default short getNumber() {
+    default Integer getNumber() {
         return 1;
     }
 }
